requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
cattrs==23.2.3
    # via jsii
cdk8s==2.68.20
    # via cdk8s-utils (pyproject.toml)
constructs==10.3.0
    # via cdk8s
importlib-resources==6.1.1
    # via jsii
jsii==1.93.0
    # via
    #   cdk8s
    #   constructs
publication==0.0.3
    # via
    #   cdk8s
    #   constructs
    #   jsii
python-dateutil==2.8.2
    # via jsii
ruamel-yaml==0.18.5
    # via cdk8s-utils (pyproject.toml)
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   cdk8s
    #   constructs
    #   jsii
typing-extensions==4.9.0
    # via jsii
